refactor: change directory for app
