+++ class-components/src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 
-import { App } from "./App.tsx";
+import { App } from "./app/App.tsx";
 
 import "./global.css";
